requirements.txt
dj-database-url==0.4.2
Django==3.2.13
Markdown==2.6.9
python-decouple==3.1
requests==2.26.0
django-widgets-improved
django-crispy-forms
bleach==3.2.1
numpy
scipy
matplotlib
ipython
jupyter
MDAnalysis
pandas
sympy
nose
biopython
docutils==0.17.1
celery
redis
redis-server
